Move Smartphone Sensor Capturer to its own file
